Moved tests into Model folder
